SoapySDR High Level API

SoapySDR.Device — Type

Device

A device is a collection of SDR channels, obtained from the Devices() list.

Fields:

info

driver

hardware

hardwareinfo

tx

rx

SoapySDR.Devices — Type

Devices()

Enumerates all detectable SDR devices on the system. Indexing into this list return a Device struct.

SoapySDR.SampleBuffer — Type

SampleBuffer(s::SoapySDR.Stream)
SampleBuffer(s::SoapySDR.Stream, n::Int)

Constructs a sample buffer for a given stream. Can contain multiple channels and be of arbitrary length. To avoid undefined behavior, this requested length with be aligned to the device MTU. It is therefore important to ensure that subsequent calls and calculations use this length.

Returns a SampleBuffer{N,T} with fields: bufs::NTuple{N, T} packet_count::Int timens::Vector{Pair{Int,typeof(1u"ns")}}

where N is the number of channels and T is the vector type of the buffer (default: Vector).

bufs are the buffers for each channel. length length of the buffer. packet_count are the number of transactions of MTU size required by subsequent read and write operations. timens are the offset and time stamp pairs for each packet.

Base.read! — Method

read!(s::SoapySDR.Stream, buf::SampleBuffer; timeout::Int)

Read data from the device into the given buffer.

SoapySDR.get_hardware_time — Method

get_hardware_time(::Device, what::String)

Get hardware time for the given source.

SoapySDR.get_sensor_info — Method

get_sensor_info(::Device, ::String)

Read the sensor extracted from list_sensors. Returns: the value as a string. Note: Appropriate conversions need to be done by the user.

SoapySDR.get_time_source — Method

get_time_source(::Device)

Set the time source used by the Device.

SoapySDR.has_hardware_time — Method

has_hardware_time(::Device, what::String)

Query if the Device has hardware time for the given source.

SoapySDR.list_sample_rates — Method

list_sample_rates(::Channel)

List the natively supported sample rates for a given channel.

SoapySDR.list_sensors — Method

list_sensors(::Device)

List the available sensors on a device. Returns: an array of sensor names.

SoapySDR.list_time_sources — Method

list_time_sources(::Device)

List time sources available on the device

SoapySDR.native_stream_format — Method

native_stream_format(c::Channel) -> Type, fullscale

Returns the format type and fullscale resolution of the native stream.

SoapySDR.read_sensor — Method

read_sensor(::Device, ::String)

Read the sensor extracted from list_sensors. Returns: the value as a string. Note: Appropriate conversions need to be done by the user.

SoapySDR.set_hardware_time — Method

has_hardware_time(::Device, timeNs::Int64 what::String)

Set hardware time for the given source.

SoapySDR.set_time_source! — Method

set_time_source!(::Device, source::String)

List the current time source used by the Device.

SoapySDR.stream_formats — Method

stream_formats(::Channel)

Returns the stream formats supported by the device.

Note: Since Julia is a multiple dispatch and generic language, it is preferrable to use native_stream_format(c::Channel) for optimal processing throughput. Only use this function if non-standard formats such as Complex Int12 and Complex Int4 are native to the device and not handled by dispatch on Complex.