SoapySDR Low Level API

SoapySDR.SoapySDRDevice_getChannelInfo — Method

SoapySDRDevice_getChannelInfo(device, direction, channel)

Get channel info given the streaming direction

device a pointer to a device instance

direction the channel direction RX or TX

channel the channel number to get info for

Returns channel information

NOTE: This function is part of the lowlevel libsoapysdr interface. For end-users in Julia, the higher-level Julia APIs are preferred

SoapySDR.SoapySDRDevice_getFrequencyRange — Method

Get the range of overall frequency values.

device a pointer to a device instance

direction the channel direction RX or TX

channel the channel number to get info for

Returns a list of frequency ranges in Hz

NOTE: This function is part of the lowlevel libsoapysdr interface. For end-users in Julia, the higher-level Julia APIs are preferred

SoapySDR.SoapySDRDevice_getFrequencyRangeComponent — Method

Get the range of tunable values for the specified element.

device a pointer to a device instance

direction the channel direction RX or TX

channel the channel number to get info for

Returns a list of frequency ranges in Hz

NOTE: This function is part of the lowlevel libsoapysdr interface. For end-users in Julia, the higher-level Julia APIs are preferred

SoapySDR.SoapySDRDevice_getGainElementRange — Method

SoapySDRDevice_getGainElementRange(device, direction, channel, name)

Get the range of possible gain values for a specific element.

device a pointer to a device instance

direction the channel direction RX or TX

channel the channel number to get info for

name the name of an amplification element

Returns the range of possible gain values for the specified amplification element in dB

NOTE: This function is part of the lowlevel libsoapysdr interface. For end-users in Julia, the higher-level Julia APIs are preferred

SoapySDR.SoapySDRDevice_getGainRange — Method

Get the overall range of possible gain values

device a pointer to a device instance

direction the channel direction RX or TX

channel the channel number to get info for

Returns the range of possible gain values for this channel in dB

SoapySDR.SoapySDRDevice_getSampleRateRange — Method

Get the range of possible baseband sample rates.

device a pointer to a device instance

direction the channel direction RX or TX

channel the channel number to get info for

Returns a list of samples rates in samples per second

NOTE: This function is part of the lowlevel libsoapysdr interface. For end-users in Julia, the higher-level Julia APIs are preferred

SoapySDR.SoapySDRDevice_getStreamFormats — Method

Query a list of the available stream formats.

device a pointer to a device instance

direction the channel direction RX or TX

channel the channel number to get info for

Returns a list of allowed format strings.

NOTE: This function is part of the lowlevel libsoapysdr interface. For end-users in Julia, the higher-level Julia APIs are preferred

SoapySDR.SoapySDRDevice_lastStatus — Method

SoapySDRDevice_lastStatus()

Get the last status code after a Device API call. The status code is cleared on entry to each Device call. When an device API call throws, the C bindings catch the exception, and set a non-zero last status code. Use lastStatus() to determine success/failure for Device calls without integer status return codes.

NOTE: This function is part of the lowlevel libsoapysdr interface. For end-users in Julia, the higher-level Julia APIs are preferred

SoapySDR.SoapySDRDevice_listAntennas — Method

Get a list of available antennas to select on a given chain.

device a pointer to a device instance

direction the channel direction RX or TX

channel the channel number to get info for

returns a list of available antenna names

NOTE: This function is part of the lowlevel libsoapysdr interface. For end-users in Julia, the higher-level Julia APIs are preferred

SoapySDR.SoapySDRDevice_listFrequencies — Method

List available tunable elements in the chain. Elements should be in order RF to baseband.

device a pointer to a device instance

direction the channel direction RX or TX

channel the channel number to get info for

Returns a list of tunable elements by name

NOTE: This function is part of the lowlevel libsoapysdr interface. For end-users in Julia, the higher-level Julia APIs are preferred

*/

SoapySDR.SoapySDRDevice_make — Method

SoapySDRDevice_make(args)

Make a new Device object given device construction args. The device pointer will be stored in a table so subsequent calls with the same arguments will produce the same device. For every call to make, there should be a matched call to unmake. param args device construction key/value argument map return a pointer to a new Device object.

NOTE: This function is part of the lowlevel libsoapysdr interface. For end-users in Julia, the higher-level Julia APIs are preferred

SoapySDR.SoapySDRDevice_setupStream — Method

Initialize a stream given a list of channels and stream arguments. The implementation may change switches or power-up components. All stream API calls should be usable with the new stream object after setupStream() is complete, regardless of the activity state.

The API allows any number of simultaneous TX and RX streams, but many dual-channel devices are limited to one stream in each direction, using either one or both channels. This call will return an error if an unsupported combination is requested, or if a requested channel in this direction is already in use by another stream.

When multiple channels are added to a stream, they are typically expected to have the same sample rate. See SoapySDRDevice_setSampleRate().

device a pointer to a device instance

direction the channel direction (SOAPY_SDR_RX or SOAPY_SDR_TX)

format A string representing the desired buffer format in read/writeStream()

The first character selects the number type:

"C" means complex

"F" means floating point

"S" means signed integer

"U" means unsigned integer

The type character is followed by the number of bits per number (complex is 2x this size per sample)

Example format strings:

"CF32" - complex float32 (8 bytes per element)

"CS16" - complex int16 (4 bytes per element)

"CS12" - complex int12 (3 bytes per element)

"CS4" - complex int4 (1 byte per element)

"S32" - int32 (4 bytes per element)

"U8" - uint8 (1 byte per element)

channels a list of channels or empty for automatic

numChans the number of elements in the channels array

args stream args or empty for defaults

Recommended keys to use in the args dictionary: - "WIRE" - format of the samples between device and host

Returns the opaque pointer to a stream handle.

NOTE:

The returned stream is not required to have internal locking, and may not be used
concurrently from multiple threads.

SoapySDR.SoapySDRDevice_writeStream — Method

Write elements to a stream for transmission. This is a multi-channel call, and buffs should be an array of void *, where each pointer will be filled with data for a different channel.

Client code compatibility: Client code relies on writeStream() for proper back-pressure. The writeStream() implementation must enforce the timeout such that the call blocks until space becomes available or timeout expiration.

device a pointer to a device instance

stream the opaque pointer to a stream handle

buffs an array of void* buffers num chans in size

numElems the number of elements in each buffer

flags optional input flags and output flags

timeNs the buffer's timestamp in nanoseconds

timeoutUs the timeout in microseconds

Returns the number of elements written per buffer, output flags