SoapySDR High Level API

SoapySDR.Device — Type

Device

A device is a collection of SDR channels, obtained from the Devices() list.

Fields:

info

driver

hardware

hardwareinfo

tx

rx

SoapySDR.Devices — Type

Devices()

Enumerates all detectable SDR devices on the system. Indexing into this list return a Device struct.

Base.read — Method

read(s::SoapySDR.Stream, nb::Integer; all=true)

Read at most nb bytes from s, returning a SampleBuffer

If all is true (the default), this function will block repeatedly trying to read all requested bytes, until an error or end-of-file occurs. If all is false, at most one read call is performed, and the amount of data returned is device-dependent. Note that not all stream types support the all option.

SoapySDR.get_hardware_time — Method

get_hardware_time(::Device, what::String)

Get hardware time for the given source.

SoapySDR.get_sensor_info — Method

get_sensor_info(::Device, ::String)

Read the sensor extracted from list_sensors. Returns: the value as a string. Note: Appropriate conversions need to be done by the user.

SoapySDR.get_time_source — Method

get_time_source(::Device)

Set the time source used by the Device.

SoapySDR.has_hardware_time — Method

has_hardware_time(::Device, what::String)

Query if the Device has hardware time for the given source.

SoapySDR.list_sample_rates — Method

list_sample_rates(::Channel)

List the natively supported sample rates for a given channel.

SoapySDR.list_sensors — Method

list_sensors(::Device)

List the available sensors on a device. Returns: an array of sensor names.

SoapySDR.list_time_sources — Method

list_time_sources(::Device)

List time sources available on the device

SoapySDR.native_stream_format — Method

native_stream_format(c::Channel) -> Type, fullscale

Returns the format type and fullscale resolution of the native stream.

SoapySDR.read_sensor — Method

read_sensor(::Device, ::String)

Read the sensor extracted from list_sensors. Returns: the value as a string. Note: Appropriate conversions need to be done by the user.

SoapySDR.set_hardware_time — Method

has_hardware_time(::Device, timeNs::Int64 what::String)

Set hardware time for the given source.

SoapySDR.set_time_source! — Method

set_time_source!(::Device, source::String)

List the current time source used by the Device.

SoapySDR.stream_formats — Method

stream_formats(::Channel)

Returns the stream formats supported by the device.

Note: Since Julia is a multiple dispatch and generic language, it is preferrable to use native_stream_format(c::Channel) for optimal processing throughput. Only use this function if non-standard formats such as Complex Int12 and Complex Int4 are native to the device and not handled by dispatch on Complex.