SoapySDR High Level API

SoapySDR.Channel — Type

Channel

A channel on the given Device.

Note!!: A Channel can be created from a Device or extracted from a ChannelList. It should rarely be necessary to create a Channel directly.

Has the following properties:

device::Device - Device to which the Channel belongs

direction - either Tx or Rx

idx - channel index used by Soapy

info - channel info consiting of OwnedKWArgs

antenna - antenna name

gain_mode - Automatic Gain control, true, false, or missing

gain_elements - list of GainElements of the channel

gain - effective gain, distributed amongst the GainElements

dc_offset_mode - Automatic DC offset mode, true, false or missing

dc_offset - DC offset value

iq_balance_mode - Automatic IQ balance mode, true, false or missing

iq_balance - IQ balance value

frequency_correction - frequency correction value

sample_rate - sample rate

bandwidth - bandwidth

frequency - center frequency

fullduplex - full duplex mode with other (TX/RX) channels

native_stream_format - native stream format

stream_formats - supported stream formats (converted by Soapy)

fullscale - full scale value

sensors - sensor list

Reading and writing to Components

gains, antennas, and sensors may consist of a chain or selectable subcomponets. To set or read e.g. a sensors, one may use the following syntax:

dev = Devices()[1] cr = dev.rx[1]

read a sensor value

s1 = cr.sensors[1] cr[s1]

read and set the gain element

g1 = cr.gain_elements[1] cr[g1] cr[g1] = 4*u"dB"

SoapySDR.ChannelList — Type

ChannelList

A grouping of channels on the Device. Note: This should not be called directly, but rather through the Device.rx and Device.tx properties.

SoapySDR.Device — Type

Device

A device is a collection of SDR channels, obtained from the Devices() list.

Fields:

info

driver

hardware

tx

rx

sensors

time_source

time_sources

frontendmapping_rx

frontendmapping_tx

SoapySDR.Devices — Type

Devices()

Enumerates all detectable SDR devices on the system. Indexing into the returned Devices object returns a list of keywords used to create a Device struct.

SoapySDR.Stream — Type

SoapySDR.Stream(channels)
SoapySDR.Stream(::Type{T}, channels)

Constructs a Stream{T} where T is the stream type of the device. If unspecified, the native format will be used.

Fields:

nchannels - The

mtu - The stream Maximum Transmission Unit

Example

SoapySDR.Stream(Devices()[1].rx)

Base.read! — Method

read!(s::SoapySDR.Stream{T}, buffer::NTuple{N, Vector{T}}; [timeout])

Read data from the device into the given buffer.

Base.read — Method

read(s::SoapySDR.Stream{T}, nb::Integer; [timeout])

Read at most nb samples from s

Base.write — Method

write(s::SoapySDR.Stream{T}, buffer::NTuple{N, Vector{T}}; [timeout]) where {N, T}

Write data from the device into the given buffer.

SoapySDR.get_hardware_time — Method

get_hardware_time(::Device, what::String)

Get hardware time for the given source.

SoapySDR.get_sensor_info — Method

get_sensor_info(::Device, ::String)

Read the sensor extracted from list_sensors. Returns: the value as a string. Note: Appropriate conversions need to be done by the user.

SoapySDR.has_hardware_time — Method

has_hardware_time(::Device, what::String)

Query if the Device has hardware time for the given source.

SoapySDR.list_sample_rates — Method

list_sample_rates(::Channel)

List the natively supported sample rates for a given channel.

SoapySDR.set_hardware_time — Method

set_hardware_time(::Device, timeNs::Int64 what::String)

Set hardware time for the given source.