SoapySDR Low Level API

Warning

This documentation is part of the lowlevel libsoapysdr interface. For end-users, the high-level Julia APIs are preferred

SoapySDR.SoapySDRDevice_acquireReadBuffer — Function

Acquire direct buffers from a receive stream. This call is part of the direct buffer access API.

The buffs array will be filled with a stream pointer for each channel. Each pointer can be read up to the number of return value elements.

The handle will be set by the implementation so that the caller may later release access to the buffers with releaseReadBuffer(). Handle represents an index into the internal scatter/gather table such that handle is between 0 and num direct buffers - 1.

param device a pointer to a device instance param stream the opaque pointer to a stream handle param handle an index value used in the release() call param buffs an array of void* buffers num chans in size param flags optional flag indicators about the result param timeNs the buffer's timestamp in nanoseconds param timeoutUs the timeout in microseconds return the number of elements read per buffer or error code

SoapySDR.SoapySDRDevice_acquireWriteBuffer — Function

Acquire direct buffers from a transmit stream. This call is part of the direct buffer access API.

The buffs array will be filled with a stream pointer for each channel. Each pointer can be written up to the number of return value elements.

The handle will be set by the implementation so that the caller may later release access to the buffers with releaseWriteBuffer(). Handle represents an index into the internal scatter/gather table such that handle is between 0 and num direct buffers - 1.

param device a pointer to a device instance param stream the opaque pointer to a stream handle param handle an index value used in the release() call param buffs an array of void* buffers num chans in size param timeoutUs the timeout in microseconds return the number of available elements per buffer or error

SoapySDR.SoapySDRDevice_activateStream — Method

Activate a stream. Call activate to prepare a stream before using read/write(). The implementation control switches or stimulate data flow.

The timeNs is only valid when the flags have SOAPYSDRHASTIME. The numElems count can be used to request a finite burst size. The SOAPYSDRENDBURST flag can signal end on the finite burst. Not all implementations will support the full range of options. In this case, the implementation returns SOAPYSDRNOT_SUPPORTED.

param device a pointer to a device instance param stream the opaque pointer to a stream handle param flags optional flag indicators about the stream param timeNs optional activation time in nanoseconds param numElems optional element count for burst control return 0 for success or error code on failure

SoapySDR.SoapySDRDevice_closeStream — Method

Close an open stream created by setupStream param device a pointer to a device instance param stream the opaque pointer to a stream handle return 0 for success or error code on failure

SoapySDR.SoapySDRDevice_deactivateStream — Method

Deactivate a stream. Call deactivate when not using using read/write(). The implementation control switches or halt data flow.

The timeNs is only valid when the flags have SOAPYSDRHASTIME. Not all implementations will support the full range of options. In this case, the implementation returns SOAPYSDRNOTSUPPORTED.

param device a pointer to a device instance param stream the opaque pointer to a stream handle param flags optional flag indicators about the stream param timeNs optional deactivation time in nanoseconds return 0 for success or error code on failure

SoapySDR.SoapySDRDevice_enumerate — Method

Enumerate a list of available devices on the system. param args device construction key/value argument filters param [out] length the number of elements in the result. return a list of arguments strings, each unique to a device

SoapySDR.SoapySDRDevice_enumerateStrArgs — Method

Enumerate a list of available devices on the system. Markup format for args: "keyA=valA, keyB=valB". param args a markup string of key/value argument filters param [out] length the number of elements in the result. return a list of arguments strings, each unique to a device

SoapySDR.SoapySDRDevice_getChannelInfo — Method

SoapySDRDevice_getChannelInfo(device, direction, channel)

Get channel info given the streaming direction device a pointer to a device instance direction the channel direction RX or TX channel the channel number to get info for

Returns channel information

NOTE: This function is part of the lowlevel libsoapysdr interface. For end-users in Julia, the higher-level Julia APIs are preferred

SoapySDR.SoapySDRDevice_getChannelSensorInfo — Method

Get meta-information about a channel sensor. Example: displayable name, type, range.

param device a pointer to a device instance param direction the channel direction RX or TX param channel an available channel on the device param key the ID name of an available sensor return meta-information about a sensor

SoapySDR.SoapySDRDevice_getChannelSettingInfo — Method

Describe the allowed keys and values used for channel settings.

param device a pointer to a device instance param direction the channel direction RX or TX param channel an available channel on the device param [out] length the number of sensor names return a list of argument info structures

SoapySDR.SoapySDRDevice_getClockSource — Method

Get the clock source of the device

param device a pointer to a device instance return the name of a clock source

SoapySDR.SoapySDRDevice_getDirectAccessBufferAddrs — Method

Get the buffer addresses for a scatter/gather table entry. When the underlying DMA implementation uses scatter/gather then this call provides the user addresses for that table.

Example: The caller may query the DMA memory addresses once after stream creation to pre-allocate a re-usable ring-buffer.

param device a pointer to a device instance param stream the opaque pointer to a stream handle param handle an index value between 0 and num direct buffers - 1 param buffs an array of void* buffers num chans in size return 0 for success or error code when not supported

SoapySDR.SoapySDRDevice_getDriverKey — Method

A key that uniquely identifies the device driver. This key identifies the underlying implementation. Serveral variants of a product may share a driver. param device a pointer to a device instance

SoapySDR.SoapySDRDevice_getFrequencyRange — Method

Get the range of overall frequency values.

device a pointer to a device instance direction the channel direction RX or TX channel the channel number to get info for

Returns a list of frequency ranges in Hz

NOTE: This function is part of the lowlevel libsoapysdr interface. For end-users in Julia, the higher-level Julia APIs are preferred

SoapySDR.SoapySDRDevice_getFrequencyRangeComponent — Method

Get the range of tunable values for the specified element.

device a pointer to a device instance direction the channel direction RX or TX channel the channel number to get info for

Returns a list of frequency ranges in Hz

NOTE: This function is part of the lowlevel libsoapysdr interface. For end-users in Julia, the higher-level Julia APIs are preferred

SoapySDR.SoapySDRDevice_getFrontendMapping — Method

Get the mapping configuration string.

param device a pointer to a device instance param direction the channel direction RX or TX return the vendor-specific mapping string

SoapySDR.SoapySDRDevice_getFullDuplex — Method

Find out if the specified channel is full or half duplex.

param device a pointer to a device instance param direction the channel direction RX or TX param channel an available channel on the device return true for full duplex, false for half duplex

SoapySDR.SoapySDRDevice_getGainElementRange — Method

SoapySDRDevice_getGainElementRange(device, direction, channel, name)

Get the range of possible gain values for a specific element.

device a pointer to a device instance direction the channel direction RX or TX channel the channel number to get info for

name the name of an amplification element

Returns the range of possible gain values for the specified amplification element in dB

NOTE: This function is part of the lowlevel libsoapysdr interface. For end-users in Julia, the higher-level Julia APIs are preferred

SoapySDR.SoapySDRDevice_getGainRange — Method

Get the overall range of possible gain values

device a pointer to a device instance direction the channel direction RX or TX channel the channel number to get info for

Returns the range of possible gain values for this channel in dB

SoapySDR.SoapySDRDevice_getHardwareInfo — Method

Query a dictionary of available device information. This dictionary can any number of values like vendor name, product name, revisions, serials... This information can be displayed to the user to help identify the instantiated device.

param device a pointer to a device instance

SoapySDR.SoapySDRDevice_getHardwareKey — Method

A key that uniquely identifies the hardware. This key should be meaningful to the user to optimize for the underlying hardware. param device a pointer to a device instance

SoapySDR.SoapySDRDevice_getHardwareTime — Method

Read the time from the hardware clock on the device. The what argument can refer to a specific time counter.

param device a pointer to a device instance param what optional argument return the time in nanoseconds

SoapySDR.SoapySDRDevice_getMasterClockRate — Method

Get the master clock rate of the device.

param device a pointer to a device instance return the clock rate in Hz

SoapySDR.SoapySDRDevice_getMasterClockRates — Method

Get the range of available master clock rates.

param device a pointer to a device instance param [out] length the number of ranges return a list of clock rate ranges in Hz

SoapySDR.SoapySDRDevice_getNativeDeviceHandle — Method

A handle to the native device used by the driver. The implementation may return a null value if it does not support or does not wish to provide access to the native handle.

param: device a pointer to a device instance return: a handle to the native device or null

SoapySDR.SoapySDRDevice_getNativeStreamFormat — Method

Get the hardware's native stream format for this channel. This is the format used by the underlying transport layer, and the direct buffer access API calls (when available).

param device a pointer to a device instance param direction the channel direction RX or TX param channel an available channel on the device param [out] fullScale the maximum possible value return the native stream buffer format string

SoapySDR.SoapySDRDevice_getNumChannels — Method

Get a number of channels given the streaming direction

param device a pointer to a device instance param direction the channel direction RX or TX return the number of channels

SoapySDR.SoapySDRDevice_getNumDirectAccessBuffers — Method

How many direct access buffers can the stream provide? This is the number of times the user can call acquire() on a stream without making subsequent calls to release(). A return value of 0 means that direct access is not supported.

param device a pointer to a device instance param stream the opaque pointer to a stream handle return the number of direct access buffers or 0

SoapySDR.SoapySDRDevice_getReferenceClockRate — Method

Get the reference clock rate of the device.

param device a pointer to a device instance return the clock rate in Hz

SoapySDR.SoapySDRDevice_getReferenceClockRates — Method

Get the range of available reference clock rates.

param device a pointer to a device instance

return a list of clock rate ranges in Hz

SoapySDR.SoapySDRDevice_getSampleRateRange — Method

Get the range of possible baseband sample rates.

device a pointer to a device instance direction the channel direction RX or TX channel the channel number to get info for

Returns a list of samples rates in samples per second

NOTE: This function is part of the lowlevel libsoapysdr interface. For end-users in Julia, the higher-level Julia APIs are preferred

SoapySDR.SoapySDRDevice_getSensorInfo — Method

Get meta-information about a sensor. Example: displayable name, type, range.

param device a pointer to a device instance param key the ID name of an available sensor return meta-information about a sensor

SoapySDR.SoapySDRDevice_getSettingInfo — Method

Describe the allowed keys and values used for settings.

param device a pointer to a device instance param [out] length the number of sensor names return a list of argument info structures

SoapySDR.SoapySDRDevice_getStreamArgsInfo — Method

Query the argument info description for stream args.

param device a pointer to a device instance param direction the channel direction RX or TX param channel an available channel on the device param [out] length the number of argument infos return a list of argument info structures

SoapySDR.SoapySDRDevice_getStreamFormats — Method

Query a list of the available stream formats.

device a pointer to a device instance direction the channel direction RX or TX channel the channel number to get info for

Returns a list of allowed format strings.

NOTE: This function is part of the lowlevel libsoapysdr interface. For end-users in Julia, the higher-level Julia APIs are preferred

SoapySDR.SoapySDRDevice_getStreamMTU — Method

Get the stream's maximum transmission unit (MTU) in number of elements. The MTU specifies the maximum payload transfer in a stream operation. This value can be used as a stream buffer allocation size that can best optimize throughput given the underlying stream implementation.

param device a pointer to a device instance param stream the opaque pointer to a stream handle return the MTU in number of stream elements (never zero)

SoapySDR.SoapySDRDevice_getTimeSource — Method

Get the time source of the device

param device a pointer to a device instance return the name of a time source

SoapySDR.SoapySDRDevice_hasHardwareTime — Method

Does this device have a hardware clock?

param device a pointer to a device instance param what optional argument return true if the hardware clock exists

SoapySDR.SoapySDRDevice_lastError — Method

Get the last error message after a device call fails. When an device API call throws, the C bindings catch the exception, store its message in thread-safe storage, and return a non-zero status code to indicate failure. Use lastError() to access the exception's error message.

SoapySDR.SoapySDRDevice_lastStatus — Method

SoapySDRDevice_lastStatus()

Get the last status code after a Device API call. The status code is cleared on entry to each Device call. When an device API call throws, the C bindings catch the exception, and set a non-zero last status code. Use lastStatus() to determine success/failure for Device calls without integer status return codes.

NOTE: This function is part of the lowlevel libsoapysdr interface. For end-users in Julia, the higher-level Julia APIs are preferred

SoapySDR.SoapySDRDevice_listAntennas — Method

Get a list of available antennas to select on a given chain.

device a pointer to a device instance direction the channel direction RX or TX channel the channel number to get info for

returns a list of available antenna names

NOTE: This function is part of the lowlevel libsoapysdr interface. For end-users in Julia, the higher-level Julia APIs are preferred

SoapySDR.SoapySDRDevice_listChannelSensors — Method

List the available channel readback sensors. A sensor can represent a reference lock, RSSI, temperature.

param device a pointer to a device instance param direction the channel direction RX or TX param channel an available channel on the device

return a list of available sensor string names

SoapySDR.SoapySDRDevice_listClockSources — Method

Get the list of available clock sources.

param device a pointer to a device instance

return a list of clock source names

SoapySDR.SoapySDRDevice_listFrequencies — Method

List available tunable elements in the chain. Elements should be in order RF to baseband.

device a pointer to a device instance direction the channel direction RX or TX channel the channel number to get info for

Returns a list of tunable elements by name

NOTE: This function is part of the lowlevel libsoapysdr interface. For end-users in Julia, the higher-level Julia APIs are preferred

*/

SoapySDR.SoapySDRDevice_listGPIOBanks — Method

Get a list of available GPIO banks by name.

param [out] length the number of GPIO banks param device a pointer to a device instance

SoapySDR.SoapySDRDevice_listGains — Method

List available amplification elements. Elements should be in order RF to baseband. param: device a pointer to a device instance param: direction the channel direction RX or TX param: channel an available channel param: [out] length the number of gain names

return a list of gain string names

SoapySDR.SoapySDRDevice_listRegisterInterfaces — Method

Get a list of available register interfaces by name.

param device a pointer to a device instance param [out] length the number of interfaces return a list of available register interfaces

SoapySDR.SoapySDRDevice_listSampleRates — Method

Get the list of possible baseband sample rates.

device a pointer to a device instance direction the channel direction RX or TX channel the channel number to get info for

Returns a list of samples rates in samples per second

NOTE: This function is part of the lowlevel libsoapysdr interface. For end-users in Julia, the higher-level Julia APIs are preferred

SoapySDR.SoapySDRDevice_listSensors — Method

List the available global readback sensors. A sensor can represent a reference lock, RSSI, temperature.

param device a pointer to a device instance param [out] length the number of sensor names return a list of available sensor string names

SoapySDR.SoapySDRDevice_listTimeSources — Method

Get the list of available time sources.

param device a pointer to a device instance param [out] length the number of sources return a list of time source names

SoapySDR.SoapySDRDevice_listUARTs — Method

Enumerate the available UART devices.

param device a pointer to a device instance param [out] length the number of UART names return a list of names of available UARTs

SoapySDR.SoapySDRDevice_make — Method

SoapySDRDevice_make(args)

Make a new Device object given device construction args. The device pointer will be stored in a table so subsequent calls with the same arguments will produce the same device. For every call to make, there should be a matched call to unmake. param args device construction key/value argument map return a pointer to a new Device object.

NOTE: This function is part of the lowlevel libsoapysdr interface. For end-users in Julia, the higher-level Julia APIs are preferred

SoapySDR.SoapySDRDevice_makeStrArgs — Method

Make a new Device object given device construction args. The device pointer will be stored in a table so subsequent calls with the same arguments will produce the same device. For every call to make, there should be a matched call to unmake.

param args a markup string of key/value arguments return a pointer to a new Device object or null for error

SoapySDR.SoapySDRDevice_make_list — Method

Create a list of devices from a list of construction arguments. This is a convenience call to parallelize device construction, and is fundamentally a parallel for loop of make(Kwargs). param argsList a list of device arguments per each device param length the length of the argsList array return a list of device pointers per each specified argument

SoapySDR.SoapySDRDevice_readChannelSensor — Method

Readback a channel sensor given the name. The value returned is a string which can represent a boolean ("true"/"false"), an integer, or float.

param device a pointer to a device instance param direction the channel direction RX or TX param channel an available channel on the device param key the ID name of an available sensor return the current value of the sensor

SoapySDR.SoapySDRDevice_readChannelSetting — Method

Read an arbitrary channel setting on the device.

param device a pointer to a device instance param direction the channel direction RX or TX param channel an available channel on the device param key the setting identifier return the setting value

SoapySDR.SoapySDRDevice_readGPIO — Method

Readback the value of a GPIO bank.

param device a pointer to a device instance param bank the name of an available bank return an integer representing GPIO bits

SoapySDR.SoapySDRDevice_readGPIODir — Method

Read the data direction of a GPIO bank.

param device a pointer to a device instance 1 bits represent outputs, 0 bits represent inputs. param bank the name of an available bank return an integer representing data direction bits

SoapySDR.SoapySDRDevice_readI2C — Method

Read from an available I2C slave. If the device contains multiple I2C masters, the address bits can encode which master. Pass the number of bytes to be read in via numBytes; numBytes will be set to the number of actual bytes read.

param device a pointer to a device instance param addr the address of the slave param [in,out] numBytes the number of bytes to read return an array of bytes read from the slave

SoapySDR.SoapySDRDevice_readRegister — Method

Read a register on the device given the interface name.

param device a pointer to a device instance param name the name of a available register interface param addr the register address return the register value

SoapySDR.SoapySDRDevice_readRegisters — Method

Read a memory block on the device given the interface name. Pass the number of words to be read in via length; length will be set to the number of actual words read.

param device a pointer to a device instance param name the name of a available memory block interface param addr the memory block start address param [in,out] length number of words to be read from memory block return the memory block content

SoapySDR.SoapySDRDevice_readSensor — Method

Readback a global sensor given the name. The value returned is a string which can represent a boolean ("true"/"false"), an integer, or float.

param device a pointer to a device instance param key the ID name of an available sensor return the current value of the sensor

SoapySDR.SoapySDRDevice_readSetting — Method

Read an arbitrary setting on the device.

param device a pointer to a device instance param key the setting identifier return the setting value

SoapySDR.SoapySDRDevice_readStream — Method

Read elements from a stream for reception. This is a multi-channel call, and buffs should be an array of void *, where each pointer will be filled with data from a different channel.

Client code compatibility: The readStream() call should be well defined at all times, including prior to activation and after deactivation. When inactive, readStream() should implement the timeout specified by the caller and return SOAPYSDRTIMEOUT.

param device a pointer to a device instance param stream the opaque pointer to a stream handle param buffs an array of void* buffers num chans in size param numElems the number of elements in each buffer param [out] flags optional flag indicators about the result param [out] timeNs the buffer's timestamp in nanoseconds param timeoutUs the timeout in microseconds return the number of elements read per buffer or error code

SoapySDR.SoapySDRDevice_readStreamStatus — Method

Readback status information about a stream. This call is typically used on a transmit stream to report time errors, underflows, and burst completion.

Client code compatibility: Client code may continually poll readStreamStatus() in a loop. Implementations of readStreamStatus() should wait in the call for a status change event or until the timeout expiration. When stream status is not implemented on a particular stream, readStreamStatus() should return SOAPYSDRNOT_SUPPORTED. Client code may use this indication to disable a polling loop.

param device a pointer to a device instance param stream the opaque pointer to a stream handle param chanMask to which channels this status applies param flags optional input flags and output flags param timeNs the buffer's timestamp in nanoseconds param timeoutUs the timeout in microseconds return 0 for success or error code like timeout

SoapySDR.SoapySDRDevice_readUART — Method

Read bytes from a UART until timeout or newline. Its up to the implementation to set the baud rate, carriage return settings, flushing on newline.

param device a pointer to a device instance param which the name of an available UART param timeoutUs a timeout in microseconds return a null terminated array of bytes

SoapySDR.SoapySDRDevice_releaseReadBuffer — Method

Release an acquired buffer back to the receive stream. This call is part of the direct buffer access API.

param device a pointer to a device instance param stream the opaque pointer to a stream handle param handle the opaque handle from the acquire() call

SoapySDR.SoapySDRDevice_releaseWriteBuffer — Function

Release an acquired buffer back to the transmit stream. This call is part of the direct buffer access API.

Stream meta-data is provided as part of the release call, and not the acquire call so that the caller may acquire buffers without committing to the contents of the meta-data, which can be determined by the user as the buffers are filled.

param device a pointer to a device instance param stream the opaque pointer to a stream handle param handle the opaque handle from the acquire() call param numElems the number of elements written to each buffer param flags optional input flags and output flags param timeNs the buffer's timestamp in nanoseconds

SoapySDR.SoapySDRDevice_setClockSource — Method

Set the clock source on the device

param device a pointer to a device instance param source the name of a clock source return an error code or 0 for success

SoapySDR.SoapySDRDevice_setFrequency — Method

Set the center frequency of the chain.

For RX, this specifies the down-conversion frequency.

For TX, this specifies the up-conversion frequency.

The default implementation of setFrequency() will tune the "RF" component as close as possible to the requested center frequency. Tuning inaccuracies will be compensated for with the "BB" component. The args can be used to augment the tuning algorithm.

Use "OFFSET" to specify an "RF" tuning offset, usually with the intention of moving the LO out of the passband. The offset will be compensated for using the "BB" component.

Use the name of a component for the key and a frequency in Hz as the value (any format) to enforce a specific frequency. The other components will be tuned with compensation to achieve the specified overall frequency.

Use the name of a component for the key and the value "IGNORE" so that the tuning algorithm will avoid altering the component.

Vendor specific implementations can also use the same args to augment tuning in other ways such as specifying fractional vs integer N tuning.

param device a pointer to a device instance param direction the channel direction RX or TX param channel an available channel on the device param frequency the center frequency in Hz param args optional tuner arguments return an error code or 0 for success

SoapySDR.SoapySDRDevice_setFrontendMapping — Method

Set the frontend mapping of available DSP units to RF frontends. This mapping controls channel mapping and channel availability.

param device a pointer to a device instance param direction the channel direction RX or TX param mapping a vendor-specific mapping string return an error code or 0 for success

SoapySDR.SoapySDRDevice_setGainElement — Method

Set the value of a amplification element in a chain.

param device a pointer to a device instance param direction the channel direction RX or TX param channel an available channel on the device param name the name of an amplification element param value the new amplification value in dB return an error code or 0 for success

SoapySDR.SoapySDRDevice_setHardwareTime — Method

Write the time to the hardware clock on the device. The what argument can refer to a specific time counter.

param device a pointer to a device instance param timeNs time in nanoseconds param what optional argument return 0 for success or error code on failure

SoapySDR.SoapySDRDevice_setMasterClockRate — Method

Set the master clock rate of the device.

param device a pointer to a device instance param rate the clock rate in Hz return an error code or 0 for success

SoapySDR.SoapySDRDevice_setReferenceClockRate — Method

Set the reference clock rate of the device.

param device a pointer to a device instance param rate the clock rate in Hz return an error code or 0 for success

SoapySDR.SoapySDRDevice_setTimeSource — Method

Set the time source on the device

param device a pointer to a device instance param source the name of a time source return an error code or 0 for success

SoapySDR.SoapySDRDevice_setupStream — Method

Initialize a stream given a list of channels and stream arguments. The implementation may change switches or power-up components. All stream API calls should be usable with the new stream object after setupStream() is complete, regardless of the activity state.

The API allows any number of simultaneous TX and RX streams, but many dual-channel devices are limited to one stream in each direction, using either one or both channels. This call will return an error if an unsupported combination is requested, or if a requested channel in this direction is already in use by another stream.

When multiple channels are added to a stream, they are typically expected to have the same sample rate. See SoapySDRDevice_setSampleRate().

device a pointer to a device instance

direction the channel direction (SOAPY_SDR_RX or SOAPY_SDR_TX)

format A string representing the desired buffer format in read/writeStream()

The first character selects the number type:

"C" means complex

"F" means floating point

"S" means signed integer

"U" means unsigned integer

The type character is followed by the number of bits per number (complex is 2x this size per sample)

Example format strings:

"CF32" - complex float32 (8 bytes per element)

"CS16" - complex int16 (4 bytes per element)

"CS12" - complex int12 (3 bytes per element)

"CS4" - complex int4 (1 byte per element)

"S32" - int32 (4 bytes per element)

"U8" - uint8 (1 byte per element)

channels a list of channels or empty for automatic

numChans the number of elements in the channels array

args stream args or empty for defaults

Recommended keys to use in the args dictionary: - "WIRE" - format of the samples between device and host

Returns the opaque pointer to a stream handle.

NOTE:

The returned stream is not required to have internal locking, and may not be used
concurrently from multiple threads.

SoapySDR.SoapySDRDevice_transactSPI — Method

Perform a SPI transaction and return the result. Its up to the implementation to set the clock rate, and read edge, and the write edge of the SPI core. SPI slaves without a readback pin will return 0.

If the device contains multiple SPI masters, the address bits can encode which master.

param device a pointer to a device instance param addr an address of an available SPI slave param data the SPI data, numBits-1 is first out param numBits the number of bits to clock out return the readback data, numBits-1 is first in

SoapySDR.SoapySDRDevice_unmake_list — Method

Unmake or release a list of device handles and free the devices array memory as well. This is a convenience call to parallelize device destruction, and is fundamentally a parallel for loop of unmake(Device *). param devices a list of pointers to device objects param length the length of the devices array return 0 for success or error code on failure

SoapySDR.SoapySDRDevice_writeChannelSetting — Method

Write an arbitrary channel setting on the device. The interpretation is up the implementation.

param device a pointer to a device instance param direction the channel direction RX or TX param channel an available channel on the device param key the setting identifier param value the setting value return 0 for success or error code on failure

SoapySDR.SoapySDRDevice_writeGPIO — Method

Write the value of a GPIO bank.

param device a pointer to a device instance param bank the name of an available bank param value an integer representing GPIO bits return 0 for success or error code on failure

SoapySDR.SoapySDRDevice_writeGPIODir — Method

Write the data direction of a GPIO bank. 1 bits represent outputs, 0 bits represent inputs.

param device a pointer to a device instance param bank the name of an available bank param dir an integer representing data direction bits return 0 for success or error code on failure

SoapySDR.SoapySDRDevice_writeGPIODirMasked — Method

Write the data direction of a GPIO bank with modification mask. 1 bits represent outputs, 0 bits represent inputs.

param device a pointer to a device instance param bank the name of an available bank param dir an integer representing data direction bits param mask a modification mask where 1 = modify return 0 for success or error code on failure

SoapySDR.SoapySDRDevice_writeGPIOMasked — Method

Write the value of a GPIO bank with modification mask.

param device a pointer to a device instance param bank the name of an available bank param value an integer representing GPIO bits param mask a modification mask where 1 = modify return 0 for success or error code on failure

SoapySDR.SoapySDRDevice_writeI2C — Method

Write to an available I2C slave. If the device contains multiple I2C masters, the address bits can encode which master. param device a pointer to a device instance param addr the address of the slave param data an array of bytes write out param numBytes the number of bytes to write return 0 for success or error code on failure

SoapySDR.SoapySDRDevice_writeMemory — Method

Write a memory block on the device given the interface name. This can represent a memory block on a soft CPU, FPGA, IC; the interpretation is up the implementation to decide.

param device a pointer to a device instance param name the name of a available memory block interface param addr the memory block start address param value the memory block content param length the number of words in the block return 0 for success or error code on failure

SoapySDR.SoapySDRDevice_writeRegister — Method

Write a register on the device given the interface name. This can represent a register on a soft CPU, FPGA, IC; the interpretation is up the implementation to decide.

param device a pointer to a device instance param name the name of a available register interface param addr the register address param value the register value return 0 for success or error code on failure

SoapySDR.SoapySDRDevice_writeSetting — Method

Write an arbitrary setting on the device. The interpretation is up the implementation.

param device a pointer to a device instance param key the setting identifier param value the setting value return 0 for success or error code on failure

SoapySDR.SoapySDRDevice_writeStream — Method

Write elements to a stream for transmission. This is a multi-channel call, and buffs should be an array of void *, where each pointer will be filled with data for a different channel.

Client code compatibility: Client code relies on writeStream() for proper back-pressure. The writeStream() implementation must enforce the timeout such that the call blocks until space becomes available or timeout expiration.

device a pointer to a device instance

stream the opaque pointer to a stream handle

buffs an array of void* buffers num chans in size

numElems the number of elements in each buffer

flags optional input flags and output flags

timeNs the buffer's timestamp in nanoseconds

timeoutUs the timeout in microseconds

Returns the number of elements written per buffer, output flags

SoapySDR.SoapySDRDevice_writeUART — Method

Write data to a UART device. Its up to the implementation to set the baud rate, carriage return settings, flushing on newline.

param device a pointer to a device instance param which the name of an available UART param data a null terminated array of bytes return 0 for success or error code on failure